+++ R/data.R
@@ -32,7 +32,7 @@
 #' @keywords datasets
 #'
 #' @references Fischl et al., (2002). Neuron, 33:341-355
-#' (\href{https://www.ncbi.nlm.nih.gov/pubmed/11832223}{PubMed})
+#' (\href{https://www.pubmed.ncbi.nlm.nih.gov/11832223}{PubMed})
 #'
 #' @format A data.frame with 2702 observations and 9 variables
 #' \describe{